# pkg/dashboard/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Helm Dashboard API",
    "version": ""
  },
...
  "paths": {
...
    "/api/helm/repositories": {
      "get": {
        "tags": [
          "Repositories"
        ],
        "description": "Get list of Helm repositories",
        "responses": {
          "200": {
            "description": "Returns list of Helm repositories"
          }
        }
      },
      "post": {
        "tags": [
          "Repositories"
        ],
        "description": "Adds new repository",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "url"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Empty response in case repository were added"
          }
        }
      }
    },
...
    "/api/k8s/{kind}/get": {
      "parameters": [
        {
          "name": "kind",
          "in": "path",
          "description": "Kind of kubernetes resource"
        },
        {
          "name": "name",
          "in": "query",
          "description": "Name of kubernetes resource",
          "required": true
        },
        {
          "name": "namespace",
          "in": "query",
          "description": "Namespace of kubernetes resource",
          "required": true
        }
      ],
      "get": {
        "tags": [
          "K8s"
        ],
        "responses": {
          "200": {
            "description": "Returns resources information"
          }
        }
      }
    },
    "/api/k8s/{kind}/list": {
      "parameters": [
        {
          "name": "kind",
          "in": "path",
          "description": "Kind of kubernetes resource",
          "schema": {
            "enum": [
              "namespaces"
            ]
          }
        }
      ],
      "get": {
        "tags": [
          "K8s"
        ],
        "responses": {
          "200": {
            "description": "Returns list of resources"
          }
        }
      }
    },
    "/api/k8s/{kind}/describe": {
      "parameters": [
        {
          "name": "kind",
          "in": "path",
          "description": "Kind of kubernetes resource"
        },
        {
          "name": "name",
          "in": "query",
          "description": "Name of kubernetes resource",
          "required": true
        },
        {
          "name": "namespace",
          "in": "query",
          "description": "Namespace of kubernetes resource",
          "required": true
        }
      ],
      "get": {
        "tags": [
          "K8s"
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {}
            },
            "description": "Returns describe text"
          }
        }
      }
    },
...
    "/api/scanners/manifests": {
      "post": {
        "tags": [
          "Scanners"
        ],
        "description": "Scan manifests using all applicable scanners",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "manifest": {
                    "type": "string",
                    "description": "Text of manifest to scan"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Map of scan results per scanner type"
          }
        }
      }
    },
    "/api/scanners/resource/{kind}": {
      "parameters": [
        {
          "in": "path",
          "name": "kind"
        },
        {
          "in": "query",
          "name": "namespace",
          "required": true
        },
        {
          "in": "query",
          "name": "name",
          "required": true
        }
      ],
      "get": {
        "tags": [
          "Scanners"
        ],
        "description": "Scan specified k8s resource in cluster",
        "responses": {
          "200": {
            "description": "Information with scan results per scanner type"
          }
        }
      }
    },
...
  }
}